backend/requirements.txt
# ============================================================
# IMW Parking - Backend Dependencies
# Finalized Version - Mars 2026
# ============================================================

# Core Flask
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# Database
mysql-connector-python==8.2.0

# Authentication & Security
PyJWT==2.11.0
bcrypt==4.0.1

# Data Validation
marshmallow==3.21.1
email-validator==2.1.1

# Reports & Exports
openpyxl==3.1.2
reportlab==4.1.0

# Tasks & Scheduling
APScheduler==3.10.4

# Protection & Optimization
Flask-Limiter==3.5.0

# Production Deployment (Ready for Linux/Docker)
gunicorn==21.2.0

